requirements.txt
requests==2.14.2
opencv-python==3.3.0.10
numpy==1.13.1
Pillow==4.1.1
libjpeg
openjpeg
zlib